% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{both.cluster}
\alias{both.cluster}
\title{row- and column-cluster a matrix}
\usage{
both.cluster(mat, extra.row = NULL, extra.column = NULL,
  hc.method = "ward.D2")
}
\arguments{
\item{mat}{input matrix}

\item{extra.row}{extra row reordering}

\item{extra.column}{extra column reordering}

\item{hc.method}{method to use in hclust}
}
\value{
a list of clustered row, column and matrix
}
\description{
row- and column-cluster a matrix
}

