% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{CMPar}
\alias{CMPar}
\title{Color Map Parameters}
\usage{
CMPar(dmin = NULL, dmax = NULL, brewer.name = NULL, brewer.n = 3,
  colorspace.name = NULL, colorspace.n = 2, cmap = NULL,
  stop.points = NULL, grey.scale = FALSE)
}
\arguments{
\item{dmin}{minimum for continuous color map}

\item{dmax}{maximum for continuous color map}

\item{brewer.name}{palette name for RColorbrewer}

\item{brewer.n}{number of stop points in RColorbrewer for continuous color map}

\item{colorspace.name}{colorspace name}

\item{colorspace.n}{number of stops in colorspace palettes}

\item{cmap}{customized colormap name}

\item{stop.points}{custome stop points}

\item{grey.scale}{whether to use grey scale}

\item{cm}{existing color maps}
}
\value{
an object of class CMPar
}
\description{
Create color map parameters
}

