% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{ly}
\alias{ly}
\title{show layout}
\usage{
ly(x)
}
\arguments{
\item{x}{plot}
}
\description{
show layout
}
\examples{
ly(
  WHeatmap(matrix(rnorm(2000),nrow=40)) +
  WHeatmap(matrix(rnorm(2000),nrow=40), cmp=CMPar(brewer.name = 'RdBu'),
    BottomRightOf(just=c('left','top'))))
}
