% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatmap.R
\name{plot.WHeatmap}
\alias{plot.WHeatmap}
\title{plot WHeatmap}
\usage{
\method{plot}{WHeatmap}(hm, cex = 1, layout.only = FALSE,
  stand.alone = TRUE)
}
\arguments{
\item{hm}{heatmap to plot}
}
\description{
plot WHeatmap
}

