% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{print.WGroup}
\alias{print.WGroup}
\title{Draw WGroup}
\usage{
\method{print}{WGroup}(group, stand.alone = TRUE, cex = 1,
  layout.only = FALSE)
}
\arguments{
\item{group}{plot to display}

\item{cex}{for scale fonts}
}
\description{
Draw WGroup
}

