% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatmap.R
\name{print.WHeatmap}
\alias{print.WHeatmap}
\title{plot WHeatmap}
\usage{
\method{print}{WHeatmap}(hm, cex = 1, layout.only = FALSE,
  stand.alone = TRUE)
}
\arguments{
\item{hm}{an object of class WHeatmap}
}
\value{
\code{NULL}
}
\description{
plot WHeatmap
}

