% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{row.cluster}
\alias{row.cluster}
\title{row cluster a matrix}
\usage{
row.cluster(mat, ..., hc.method = "ward.D2")
}
\arguments{
\item{mat}{input matrix}

\item{...}{extra color bars or matrix that needs row reordered.}

\item{hc.method}{method to use in hclust}
}
\value{
a list of clustered row, column and matrix
}
\description{
row cluster a matrix
}

