% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{WColorBarH}
\alias{WColorBarH}
\title{WColorBarH}
\usage{
WColorBarH(data, ...)
}
\arguments{
\item{data}{numeric vector}

\item{...}{additional options to WHeatmap}
}
\value{
an object of class WColorBarH
}
\description{
a horizontal color bar
}

