% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{WColorBarV}
\alias{WColorBarV}
\title{WColorBarV}
\usage{
WColorBarV(data, ...)
}
\arguments{
\item{data}{numeric vector}

\item{...}{additional options to WHeatmap}
}
\value{
an object of class WColorBarV
}
\description{
a vertical color bar
}

