% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{WColumnBind}
\alias{WColumnBind}
\title{column bind non-overlapping objects}
\usage{
WColumnBind(..., nr = NULL, nc = NULL)
}
\arguments{
\item{...}{plotting objects}

\item{nr}{number of rows}

\item{nc}{number of columns}
}
\value{
an object of class WDim
}
\description{
column bind non-overlapping objects
}
\examples{
WHeatmap(matrix(rnorm(2000),nrow=40),name='a') +
  WHeatmap(matrix(rnorm(30), nrow=3), RightOf(),name='b') +
  WColorBarH(1:10, TopOf(WColumnBind('a','b')))
}
