% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatmap.R
\name{WHeatmap}
\alias{WHeatmap}
\title{WHeatmap object}
\usage{
WHeatmap(data = NULL, dm = NULL, name = "", continuous = NULL,
  cmp = CMPar(), cm = NULL, title = NULL, title.fontsize = 12,
  title.pad = 0.005, title.side = "l", xticklabels = NULL,
  xticklabels.n = NULL, xticklabel.side = "b", xticklabel.fontsize = 12,
  xticklabel.rotat = 90, xticklabel.pad = 0.005, yticklabels = NULL,
  yticklabels.n = NULL, yticklabel.side = "l", yticklabel.fontsize = 12,
  yticklabel.pad = 0.005, alpha = 1, sub.name = NULL, gp = NULL)
}
\arguments{
\item{data}{data matrix}

\item{dm}{plotting dimension (a WDim or a WDimGenerator object)}

\item{name}{name of the plot}

\item{continuous}{whether the data should be treated as continuous or discrete}

\item{cmp}{a CMPar object, for tunning color mapping parameters}

\item{cm}{a given color map}

\item{xticklabels}{xtick label}

\item{xticklabels.n}{number of xtick labels to plot (resample for aethetics by default)}
}
\value{
one or a list of heatmaps (depends on whether dimension is split)
}
\description{
Create a heatmap
}

