% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{both.cluster}
\alias{both.cluster}
\title{row- and column-cluster a matrix}
\usage{
both.cluster(mat, hc.method = "ward.D2")
}
\arguments{
\item{hc.method}{method to use in hclust}

\item{at}{input matrix}
}
\value{
a list of clustered row, column and matrix
}
\description{
row- and column-cluster a matrix
}

