% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{WColorBarH}
\alias{WColorBarH}
\title{WColorBarH}
\usage{
WColorBarH(
  data,
  ...,
  label = NULL,
  label.side = "r",
  label.fontsize = 12,
  label.pad = 0.005,
  label.space = 0.05,
  label.use.data = FALSE
)
}
\arguments{
\item{data}{numeric vector}

\item{...}{additional options to WHeatmap}

\item{label}{colorbar label}

\item{label.side}{l (for left) or r (for right)}

\item{label.fontsize}{label font size}

\item{label.pad}{label padding}

\item{label.space}{when label.use.data, the space between labels}

\item{label.use.data}{use data to show legend in situ}
}
\value{
an object of class WColorBarH
}
\description{
a horizontal color bar
}
\examples{
WColorBarH(matrix(1:50))
}
