% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{WColorBarV}
\alias{WColorBarV}
\title{WColorBarV}
\usage{
WColorBarV(
  data,
  ...,
  label = NULL,
  label.side = "t",
  label.fontsize = 12,
  label.pad = 0.005,
  label.space = 0.05,
  label.use.data = FALSE
)
}
\arguments{
\item{data}{numeric vector}

\item{...}{additional options to WHeatmap}

\item{label}{colorbar label}

\item{label.side}{t (for top) or b (for bottom)}

\item{label.fontsize}{label font size}

\item{label.pad}{label padding}

\item{label.space}{when label.use.data, the space between labels}

\item{label.use.data}{use data to show legend in situ}
}
\value{
an object of class WColorBarV
}
\description{
a vertical color bar
}
\examples{
WColorBarV(matrix(50:1))
}
